+++ src/model/Walklets.cpp
@@ -24,7 +24,7 @@ void Walklets::Train(int walk_times, int walk_steps, int window_min, int window_
     long total = walk_times*pnet.MAX_vid;
     double alpha_min = alpha*0.0001;
     double _alpha = alpha;
-    long count = 0;
+    unsigned long long count = 0;
 
     for (int t=0; t<walk_times; ++t)
     {
